change port to ushort

+++ src/NosCore.Shared/Configuration/ServerConfiguration.cs
@@ -14,8 +14,8 @@ namespace NosCore.Shared.Configuration
     {
         public string? Host { get; set; }
 
-        [Range(1, int.MaxValue)]
-        public int Port { get; set; }
+        [Range(1, ushort.MaxValue)]
+        public ushort Port { get; set; }
 
         public override string ToString()
         {
